Fix compilation errors: ssize to ssize_t and explicit type conversions
- Fix CStrictInteger.h: Replace 'ssize' with 'ssize_t' in concept definitions
- Fix EditInfo.h: Use brace initialization for CLayoutInt (explicit constructor)
- Fix CBregexp.h: Use brace initialization for CLogicInt (explicit constructor)
- Fix CViewCommander_Search.cpp: Add GetValue() for CLayoutInt to ssize_t conversion
- Fix CLayoutMgr.h: Add explicit static_cast<int> for Int to int conversion
- Fix DispPos.h: Add explicit static_cast<int> for Int to int conversion

+++ sakura_core/basis/CStrictInteger.h
@@ -20,11 +20,11 @@
 
 //! 整数型、または、intにキャスト可能な型
 template<typename T>
-concept IntOrCastable = std::is_integral_v<T> || std::is_convertible_v<T, int>;
+concept IntOrCastable = std::is_integral_v<T> || std::is_convertible_v<T, ssize_t>;
 
 //! 整数型、または、Int(CLazyInt)
 template<typename T>
-concept IntOrLazyInt = std::is_integral_v<T> || (!std::is_same_v<Int, int> && std::is_same_v<T, Int>);
+concept IntOrLazyInt = std::is_integral_v<T> || (!std::is_same_v<Int, ssize_t> && std::is_same_v<T, Int>);
 
 //! 暗黙の変換を許さない、整数クラス
 template <
